# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/items": {
      "post": {
        "operationId": "ItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "get": {
        "operationId": "ItemController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
    "/items/item-stocks": {
      "post": {
        "operationId": "ItemController_createItemStock",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "get": {
        "operationId": "ItemController_findAllItemStocks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
...
    "/items/item-stocks/{id}": {
      "get": {
        "operationId": "ItemController_findOneItemStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "patch": {
        "operationId": "ItemController_updateItemStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "delete": {
        "operationId": "ItemController_removeItemStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
...
    "/items/item-prices": {
      "post": {
        "operationId": "ItemController_createItemPrice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemPriceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "get": {
        "operationId": "ItemController_findAllItemPrices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
    "/items/item-prices/{id}": {
      "get": {
        "operationId": "ItemController_findOneItemPrice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "patch": {
        "operationId": "ItemController_updateItemPrice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "delete": {
        "operationId": "ItemController_removeItemPrice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
    "/items/{id}": {
      "get": {
        "operationId": "ItemController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "put": {
        "operationId": "ItemController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      },
      "delete": {
        "operationId": "ItemController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
...
    "/common": {
      "post": {
        "operationId": "CommonController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Common"
        ]
      },
      "get": {
        "operationId": "CommonController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Common"
        ]
      }
    },
    "/common/{id}": {
      "put": {
        "operationId": "CommonController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Common"
        ]
      },
      "delete": {
        "operationId": "CommonController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Common"
        ]
      }
    },
    "/common/type/{type}": {
      "get": {
        "operationId": "CommonController_getByType",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Common"
        ]
      }
    },
...
    "/business": {
      "post": {
        "operationId": "BusinessController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Business"
        ]
      },
      "get": {
        "operationId": "BusinessController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Business"
        ]
      }
    },
    "/business/{id}": {
      "get": {
        "operationId": "BusinessController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Business"
        ]
      },
      "put": {
        "operationId": "BusinessController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Business"
        ]
      },
      "delete": {
        "operationId": "BusinessController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Business"
        ]
      }
    },
...
    "/sales": {
      "post": {
        "operationId": "SaleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "sales"
        ]
      },
      "get": {
        "operationId": "SaleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sales"
        ]
      }
    },
...
    "/sales/{id}": {
      "get": {
        "operationId": "SaleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sales"
        ]
      },
      "patch": {
        "operationId": "SaleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sales"
        ]
      },
      "delete": {
        "operationId": "SaleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "Inventory System API",
    "description": "API documentation for the inventory and accounting system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateItemDto": {
        "type": "object",
        "properties": {}
      },
      "CreateItemStockDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateItemStockDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateItemPriceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateItemPriceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCommonDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCommonDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBusinessDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "number"
          },
          "itemId": {
            "type": "number"
          },
          "amountPaid": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          },
          "warehouseId": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "customerId",
          "itemId",
          "amountPaid",
          "remarks",
          "warehouseId",
          "quantity"
        ]
      },
      "UpdateSaleDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "number"
          },
          "itemId": {
            "type": "number"
          },
          "amountPaid": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          },
          "warehouseId": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          }
        }
      },
...
    }
  }
}